i3c: master: exclude controller for duplication check
In case errors on the bus or improper target device behaviour there is
a risk that PID is read as 0 or is read partially which could result
with the same. Then while duplication checking such device would be
matched with the device related to the controller because controllers
device PID equals 0. So device related to the controller shall be
excluded while duplication checking.

Signed-off-by: Zbigniew Lukwinski <zbigniew.lukwinski@linux.intel.com>
Change-Id: I8d6e6b4573e35d32a39d08cde89aaa6aad3b1e03