jtag: aspeed: Keep the TRSTn to high when jtag hw2 enable.
The TRST control bit for JTAG hardware mode 2 is part of the mode2 control
register. In order to ensure that TRST remains high when performing other
operations and to avoid potential issues with devices that require it, we
should set this bit to 1 when enabling JTAG to meet the usage of the
jtag hw mode1.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I103b05894045ad314b319e612ccc3576a470a59b