arm64: dtsi: aspeed-g7: Fix wrong use of GIC_CPU_MASK_SIMPLE() for GICv3
In 3rd cell of GICv3, there is no affinity bitmap(bits[15:8]) in the
binding(arm,gic-v3.yaml) even in PPI interrupt type.

If affinity bitmap needed, it can be specified using a 4th cell GICv3
as documented in the bindings.

Clean up the wrong use of the affinity bitmap using the
GIC_CPU_MASK_SIMPLE() macro.

Change-Id: I396975f78fd885cf93037d84a9f708e6861da8bb
Signed-off-by: Kevin Chen <kevin_chen@aspeedtech.com>