pcie: aspeed: Change to sense gpio rising edge
According to the PCIe spec, the deassertion (rising) of PERST should
indicate that the reference clock is stable. The original driver caused a
timing violation where PERST was deasserted without the reference clock
being stable. This modification also adjusts the software timing for
toggling RC PERST and the perst_rc_out GPIO.
With perst_rc_out GPIO, PERST for external devices is controlled by the
GPIO, while the internal RC PERST is connected to our EP PERST. In this
case, the RC PERST only needs to stay low for 1ms to avoid resetting the
EP PERST after the PCIe link phase is completed. However, the GPIO must
remain low for 100ms to comply with the spec.
Without perst_rc_out GPIO, the RC PERST is directly connected to external
devices, so it must follow the spec and remain low for at least 100ms.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I48a4f90cb0df4ece72b18f030a0e758b24907f75