ARM: aspeed: g5: Change the bus-frequency for sgpio
The SGPIO bus frequency can only go up to 6MHz on the AST2500 with a core
clock of 24MHz, because the clock divider of the SGPIO controller cannot
be 0. As a result, the maximum bus frequency is 24MHz / 4 = 6MHz. This
patch changes the default frequency to 1MHz to avoid the above issue.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: Ib21101add2836b0dabca2c1d3036be6afa87d379