gpio: aspeed-sgpio: Change the macro to support deferred probe
Use module_platform_driver() to replace module_platform_driver_probe().
The former utilizes platform_driver_register(), which allows the driver to
defer probing when it doesn't acquire the necessary resources due to probe
order. In contrast, the latter uses __platform_driver_probe(), which
includes the comment "Note that this is incompatible with deferred
probing." Since our SGPIO driver requires access to the clock resource,
the former is more suitable.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I3553d2f67f3727fa5f7546f415425b2b477a17aa