i3c: mipi-i3c-hci: Support priv_xfer with target mode
Add a priv_xfer API for target mode, which allows the driver to prepare
the read data and wait for the master to read it. This is different from
the pending read notification, which sends an IBI to inform the master to
read. This API can be used to implement pseudo-EEPROM functionality, which
needs to respond to the W + Sr + R command. The controller will trigger a
write-finished interrupt upon receiving the Sr signal, and the driver
should provide the data at that time. Otherwise, the hardware will issue a
NACK for the subsequent R command.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I1862fb85815a2023ec3fe528e75bf725fb4d096a