clk/rst: Change the initialization order of the clock/reset driver
This patch modifies the probe order for the clock and reset drivers:
- The clock driver is now probed at the arch initialization level.
- The reset driver is shifted to the subsys initialization level.

Since the auxiliary device has a module name, it must be initialized at
the subsys level instead of the arch level. This change is necessary
because the module_kset is initialized during the subsys stage.
Note: When the clock driver probes, it will register the reset devices.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: If8297f05d119209371d3e4bfab5e84558f1c04fd