i3c: mipi-i3c-hci: aspeed: Modify PIO register for DMA operation
Our DMA mode is affected by the PIO register settings.
The default value of the PIO register `REG_TX_START_THLD` is set to `0x1`,
which means the controller waits for at least `(2 ^ 2) DWORDs` (16 bytes)
to be written before starting transmission. As a result, in DMA mode, if
the data size is between 5 and 12 bytes, the controller does not start
transmission, causing a stall.
- Data size of 1-4 bytes: Uses Immediate Data Transfer Command.
- Data size of 13+ bytes: Since the hardware fetches data in 4-byte
  chunks, a 13-byte transfer fetches 16 bytes into the FIFO.
- Data size of 5-12 bytes: Triggers the stall due to the
  `REG_TX_START_THLD` setting.
This patch clear the `REG_TX_START_THLD` to 0 when executing with DMA mode
as workaround to resolve the above issue.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: Ib6bb57eba03a623788382920636afac7749f60a5