i3c: master: Handle the internal error in target mode only
In target mode, the controller enters the halt state only when it receives
private write requests too quickly, causing the IBI queue to become full -
which is a passive condition.
In primary mode, it enters the halt state when any transfer error occurs,
but it is automatically resumed by the transfer error IRQ handler, so
internal error handling is not required for recovery.

Fixes: 2f52d934ee69 ("i3c: master: Enhance error handling and reset mechanisms")
Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I5116ab59af11aec0a29d524e0f2e192ce8fc33be