soc: aspeed: mctp: fix rx stopped flag assign race
The stopped flag indicates whether the MCTP controller has run out of
RX CMDs for receiving MCTP packets.

The flag is set to true when the interrupt RX_CMD_NO_MORE_INT is raised.
When RX tasklet detects the flag, it set RX to ready state again,then
clear this flag to false.

A problematic scenario is that the RX tasklet may be interrupted again
by the RX_CMD_NO_MORE_INT between the RX state assignment and clearing
the stopped flag.

In the end of the IRQ, it sets the flag to true and schedule another RX
tasklet execution. Then return to the original interrupted tasklet, who
rewrites the flag to false, and misleads the next scheduled RX tasklet
execution to think the RX state is ready.

Move the stopped flag assignment before the RX ready state assignment to
avoid the next scheduled RX tasklet execution being misled.

Signed-off-by: aspeedyh <yh_chung@aspeedtech.com>
Change-Id: Ifa5ac7d8b4975915a59e2a1327e09c8e822b4d77