mctp: pcie-vdm: Fix MCTP control message parsing
Currently MCTP control message parsing code incorrectly calculates
offset of the Rq/D bits in the incoming message.
This is because the first byte of the MCTP control message is a
"Message Type" byte and the relevant structure "mctp_ctrl_msg_hdr"
is missing this field.
Add "u8 message_type" to the structure fields to fix the issue.

Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>