drivers: i2c: merge the slave IRQ handling for AST2600
In a previous patch, we introduced two new cases to handle the updated i2c
slave behavior on the AST2600. However, it turns out that both cases should
be handled using the same logic. Additionally, we discovered that one of
the cases was incorrectly invoking the virtual slave API.
This patch corrects it and merges the handling into a unified case.

Signed-off-by: Tommy Huang <tommy_huang@aspeedtech.com>
Change-Id: I1111be0b89f4836697c45f41c3cdcff5313d55b2