drivers: crypto: acry: Replace remove_new with remove
The platform_driver .remove_new callback was a temporary
alias introduced during the transition to the new .remove
prototype returning void. The kernel now uses only .remove,
and .remove_new is being deprecated.

Update the acry driver to use .remove instead of .remove_new.
This change has no functional impact and is part of the
ongoing cleanup to drop the legacy interface.

Signed-off-by: chengyu_lin <chengyu_lin@aspeedtech.com>
Change-Id: I3babc0c663789fb6957dcfcd302233189fcc7b5c