i3c: dw: Handle unexpected responses caused by SDA noise
Currently, if an IBI address appears in the DAT (including the 0x2 hot-join
address) and is marked for rejection, the controller will NACK the IBI and
automatically issue a DISCE CCC command to the device.
After performing stress tests under this condition, we observed that the
response queue may contain unexpected data.
This unexpected response always uses TID 0, so this patch reserves TID 0
and prevents it from being used for normal transactions in order to
reliably detect and handle this corner case.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I6d66b30ea02fcb3f9bc0c6514d62989ab26c65fb